Add the "explanations" folder, implement the `NOT` logic gate.
This formalizes the reason for having this repository be publicly
available. The explanations may also encourage me to more thoroughly
explain my own understanding of the material and thus prove to myself
that I actually understand the material well enough to impart the
knowledge to others in a meaningfully useful way.